requirements.txt
flask==2.3.3
flask-socketio==5.3.4
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pymongo==4.5.0
python-dateutil==2.8.2
gunicorn==21.2.0
